Issue that caused the below issue was caused by not specifying the SQL_VARCHAR buffer length.
Revert "163728 Oracle is failing to import the partition definition yml file unless DAN and JPN are removed. Its not due to an issue with these languages specifically. Both import successfully by themselves. The error message is obscure and does not describe which column is actually causing the issue. All the data in memory looks correct and so the suspicion is the Oracle driver, and yet the driver basically works for customers. Maybe with this work around further issues will manifest which may help to pinpoint the issue."

This reverts commit 00e06012090bd4f0a784830e064901aca2dc0d06.